Clone lexdata repo with HTTPS, not SSH.
Previously Lexdata was cloned using SSH. This will fail on all
new Blubber builds as it occurs within the docker image rather
than on the local machine (that may or may not have SSH access).

Also added a bit of error handling when cloning and executing
the Symbolset ss_files setup script.

Change-Id: I19dce8201dcd89dedb823673ab679b00af7a271c